% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{kimberlite_pal}
\alias{kimberlite_pal}
\title{kimberlite palette with ramped colours}
\usage{
kimberlite_pal(palette = "coolors", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'kimberlite_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
kimberlite palette with ramped colours
}
\examples{
\dontrun{
library(scales)
show_col(kimberlite_pal()(10))

filled.contour(volcano,color.palette = kimberlite_pal(), asp=1)
}
}
