% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{kimberlite_pal}
\alias{kimberlite_pal}
\title{kimberlite palette with ramped (interpolated) colours}
\usage{
kimberlite_pal(palette = "coolors", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose from 'kimberlite_palettes' list}

\item{reverse}{If TRUE, the direction of the colours is reversed.}

\item{...}{Other arguments passed to colorRampPalette()}
}
\description{
kimberlite palette with ramped (interpolated) colours
}
\examples{
\dontrun{
library(scales)
show_col(kimberlite_pal()(10))

filled.contour(volcano,color.palette = kimberlite_pal(), asp=1)
}
}
