% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_kimberlite}
\alias{scale_color_kimberlite}
\alias{scale_colour_kimberlite}
\title{Setup colour palette for ggplot2}
\usage{
scale_color_kimberlite(
  ...,
  palette = "coolors",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)

scale_colour_kimberlite(
  ...,
  palette = "coolors",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'kimberlite_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}
}
\description{
Setup colour palette for ggplot2
}
\examples{
library(kimberlite)
library(ggplot2)
ggplot(diamonds, aes(x=carat, y=price, color = clarity)) +
  geom_point() +
  scale_color_kimberlite() +
  labs(title = "default: coolors")
}
